Added option to fix a misalignment of RAW-pickles on odd image indicies
During some measurements, some weird noise was found, when using
odd values for x-coordinate of the left upper corner and/or the
the image width. This happens as the image-x coordinate is converted
into bayer-indicies by multiplying by 1.5 (12bit / 8bit). That causes
half-indicies, when having an odd index. As the index is then truncated,
a misalignment of the bayer-data appears.

FindBayerSpace-Misalignment.py can be used to determine the index of
misalignment, which then can be used by the modified _PreClipImagesOrRaw.py
to re-align the RAW-images.

Note:
That bug is already fixed in the PyCam2-Server, so that odd window-values
which are given by the user, converted into even numbers by:
ValidVal = GivenVal - (GivenVal % 2)